move BERTDatasetTransform back to script and add test for it
