+++ scripts/bert/__init__.py
@@ -19,4 +19,4 @@
 
 # pylint: disable=wildcard-import
 """BERT example."""
-from . import bert, dataset, tokenizer
+from . import bert, dataset
