+++ scripts/bert/dataset.py
@@ -17,7 +17,8 @@
 __all__ = ['MRPCDataset']
 
 import os
-from gluonnlp.data import TSVDataset
+import numpy as np
+from gluonnlp.data import TSVDataset, BERTSentenceTransform
 from gluonnlp.data.registry import register
 
 
@@ -47,3 +48,100 @@ class MRPCDataset(TSVDataset):
     def get_labels():
         """Get classification label ids of the dataset."""
         return ['0', '1']
+
+
+class BERTDatasetTransform(object):
+    """Dataset Transformation for BERT-style Sentence Classification or Regression.
+
+    Parameters
+    ----------
+    tokenizer : BERTTokenizer.
+        Tokenizer for the sentences.
+    labels : list of int or float.
+        List of all label ids for the classification task and regressing task.
+    max_seq_length : int.
+        Maximum sequence length of the sentences.
+    pad : bool, default True
+        Whether to pad the sentences to maximum length.
+    pair : bool, default True
+        Whether to transform sentences or sentence pairs.
+    label_dtype: int32 or float32, default float32
+        label_dtype = int32 for classification task
+        label_dtype = float32 for regression task
+    """
+
+    def __init__(self, tokenizer, labels, max_seq_length,
+                 pad=True, pair=True, label_dtype='float32'):
+        self.label_dtype = label_dtype
+        self._label_map = {}
+        for (i, label) in enumerate(labels):
+            self._label_map[label] = i
+        self._bert_xform = BERTSentenceTransform(tokenizer, max_seq_length, pad=pad, pair=pair)
+
+    def __call__(self, line):
+        """Perform transformation for sequence pairs or single sequences.
+
+        The transformation is processed in the following steps:
+        - tokenize the input sequences
+        - insert [CLS], [SEP] as necessary
+        - generate type ids to indicate whether a token belongs to the first
+          sequence or the second sequence.
+        - generate valid length
+
+        For sequence pairs, the input is a tuple of 3 strings:
+        text_a, text_b and label.
+
+        Inputs:
+            text_a: 'is this jacksonville ?'
+            text_b: 'no it is not'
+            label: '0'
+        Tokenization:
+            text_a: 'is this jack ##son ##ville ?'
+            text_b: 'no it is not .'
+        Processed:
+            tokens:  '[CLS] is this jack ##son ##ville ? [SEP] no it is not . [SEP]'
+            type_ids: 0     0  0    0    0     0       0 0     1  1  1  1   1 1
+            valid_length: 14
+            label: 0
+
+        For single sequences, the input is a tuple of 2 strings: text_a and label.
+        Inputs:
+            text_a: 'the dog is hairy .'
+            label: '1'
+        Tokenization:
+            text_a: 'the dog is hairy .'
+        Processed:
+            text_a:  '[CLS] the dog is hairy . [SEP]'
+            type_ids: 0     0   0   0  0     0 0
+            valid_length: 7
+            label: 1
+
+        Parameters
+        ----------
+        line: tuple of str
+            Input strings. For sequence pairs, the input is a tuple of 3 strings:
+            (text_a, text_b, label). For single sequences, the input is a tuple
+            of 2 strings: (text_a, label).
+
+        Returns
+        -------
+        np.array: input token ids in 'int32', shape (batch_size, seq_length)
+        np.array: valid length in 'int32', shape (batch_size,)
+        np.array: input token type ids in 'int32', shape (batch_size, seq_length)
+        np.array: classification task: label id in 'int32', shape (batch_size, 1),
+            regression task: label in 'float32', shape (batch_size, 1)
+        """
+        input_ids, valid_length, segment_ids = self._bert_xform(line[:-1])
+
+        label = line[-1]
+        # classification task
+        if self.label_dtype == 'int32':
+            label_id = self._label_map[label]
+            label_id = np.array([label_id], dtype='int32')
+            return input_ids, valid_length, segment_ids, label_id
+        # regression task
+        elif self.label_dtype == 'float32':
+            label = np.array([float(label)], dtype='float32')
+            return input_ids, valid_length, segment_ids, label
+        else:
+            raise ValueError('Lable dtype does not belong to [float32, int32]')
