+++ tests/unittest/test_transforms.py
@@ -259,51 +259,3 @@ def test_bert_sentences_transform():
     assert all(token_ids == valid_token_ids)
     assert length == len(vocab_tokens) + 3
     assert all(type_ids == valid_type_ids)
-
-
-def test_bert_dataset_transform():
-    text_a = 'is this jacksonville ?'
-    text_b = 'no it is not'
-    label_cls = 0
-    vocab_tokens = ['is', 'this', 'jack', '##son', '##ville', '?', 'no', 'it', 'is', 'not']
-
-    bert_vocab = BERTVocab(count_tokens(vocab_tokens))
-    tokenizer = t.BERTTokenizer(vocab=bert_vocab)
-
-    # test BERTDatasetTransform for classification task
-    bert_cls_dataset_t = t.BERTDatasetTransform(tokenizer, [label_cls], 15, pad=True,
-                                                pair=True, label_dtype='int32')
-    token_ids, length, type_ids, label_ids = bert_cls_dataset_t((text_a, text_b, label_cls))
-
-    text_a_tokens = ['is', 'this', 'jack', '##son', '##ville', '?']
-    text_b_tokens = ['no', 'it', 'is', 'not']
-    text_a_ids = bert_vocab[text_a_tokens]
-    text_b_ids = bert_vocab[text_b_tokens]
-
-    cls_ids = bert_vocab[[bert_vocab.cls_token]]
-    sep_ids = bert_vocab[[bert_vocab.sep_token]]
-    pad_ids = bert_vocab[[bert_vocab.padding_token]]
-
-    concated_ids = cls_ids + text_a_ids + sep_ids + text_b_ids + sep_ids + pad_ids
-    valid_token_ids = np.array([pad_ids[0]]*15, dtype=np.int32)
-    for i, x in enumerate(concated_ids):
-        valid_token_ids[i] = x
-    valid_type_ids = np.zeros((15,), dtype=np.int32)
-    start = len(text_a_tokens) + 2
-    end = len(text_a_tokens)+2+len(text_b_tokens)+1
-    valid_type_ids[start:end] = 1
-
-    assert all(token_ids == valid_token_ids)
-    assert length == len(vocab_tokens) + 3
-    assert all(type_ids == valid_type_ids)
-    assert all(label_ids == np.array([label_cls], dtype=np.int32))
-
-    # test BERTDatasetTransform for regression task
-    label_reg = 0.2
-    bert_reg_dataset_t = t.BERTDatasetTransform(tokenizer, [label_reg], 15, pad=True,
-                                                pair=True, label_dtype='float32')
-    token_ids, length, type_ids, label_value = bert_cls_dataset_t((text_a, text_b, label_cls))
-    assert all(token_ids == valid_token_ids)
-    assert length == len(vocab_tokens) + 3
-    assert all(type_ids == valid_type_ids)
-    assert all(label_ids == np.array([label_cls], dtype=np.int32))
